Editing post fully working, regardless of having previous image or not

+++ SocialMedia_App/Controllers/HomeController.cs
@@ -10,6 +10,8 @@ using System.Net.Http;
 using System.Web;
 using SocialMedia_App.Core.Application.Services;
 using SocialMedia_App.Core.Domain.Entities;
+using AutoMapper;
+using Microsoft.Extensions.Hosting;
 
 
 namespace SocialMedia_App.Controllers
@@ -20,14 +22,16 @@ namespace SocialMedia_App.Controllers
         private readonly ValidateUserSession _validateUserSession;
         private readonly IUserService _userService;
         private readonly FileHelper _fileHelper;
+        private readonly IMapper _mapper;
 
         public HomeController(IPostService postService, ValidateUserSession validateUserSession,
-            FileHelper fileHelper, IUserService userService)
+            FileHelper fileHelper, IUserService userService, IMapper mapper)
         {
             _postService = postService;
             _validateUserSession = validateUserSession;
             _fileHelper = fileHelper;
             _userService = userService;
+            _mapper = mapper;
         }
 
         public async Task<IActionResult> Index()
@@ -100,20 +104,75 @@ namespace SocialMedia_App.Controllers
 
         public async Task<IActionResult> Delete(int id)
         {
-           await _postService.Delete(id);
+            if (!_validateUserSession.HasUser())
+            {
+                TempData["NoAccess"] = "No tiene permiso para acceder a esta página. Primero debe iniciar sesión.";
+                return RedirectToRoute(new { controller = "Login", action = "Index" });
+            }
+
+            await _postService.Delete(id);
            return RedirectToAction("Index");
         }
 
         public async Task<IActionResult> Edit(int id)
         {
-            return RedirectToAction("Index");
+            if (!_validateUserSession.HasUser())
+            {
+                TempData["NoAccess"] = "No tiene permiso para acceder a esta página. Primero debe iniciar sesión.";
+                return RedirectToRoute(new { controller = "Login", action = "Index" });
+            }
+
+            var post = await _postService.GetByIdSaveViewModel(id);
+
+            EditPostViewModel toEdit = new();
+            toEdit.Id = post.Id;    
+            toEdit.UserId = post.UserId;
+            toEdit.Content = post.Content;
+            toEdit.YouTubeLink = post.YouTubeLink;
+            toEdit.PostImageUrl = post.PostImageUrl;
+
+            return View(toEdit);
         }
 
         [HttpPost]
-        public async Task<IActionResult> Edit()
+        public async Task<IActionResult> Edit(EditPostViewModel vm)
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "Login", action = "Index" });
+            }
+            if (!ModelState.IsValid)
+            {
+                return View(vm);
+            }
+
+            if (!string.IsNullOrWhiteSpace(vm.YouTubeLink))
+            {
+                vm.YouTubeLink = ConvertToEmbedUrl(vm.YouTubeLink);
+            }
+
+            var currentPost = await _postService.GetByIdSaveViewModel(vm.Id);
+            _mapper.Map(vm, currentPost);
+
+            await _postService.Update(currentPost, vm.Id);
+
+            if (!string.IsNullOrWhiteSpace(vm.PostImageUrl))
+            {
+                currentPost.PostImageUrl = vm.PostImageUrl;
+            }
+
+            if (vm.File != null && vm.File.Length > 0)
+            {
+                string imageUrl = _fileHelper.UploadFile(vm.File, currentPost.Id, "Posts");
+                await _postService.PostImage(currentPost.Id, imageUrl);
+            }
+            //Pero en el caso de que se esté editando un post que ya viene con una imagen originalmente,
+            //si edito un campo que no sea el de la imagen, o sea, si no subo ningun archivo, no me está
+            //conservando la imagen previa sino que se esta guardando el registro con la propiedad en null
             return RedirectToAction("Index");
         }
+
+
 
 
     }
